Update Jan 2018

+++ generate_user_files.py
@@ -7,7 +7,6 @@
 # Distributed under the terms of the MIT license.
 #
 from __future__ import absolute_import, unicode_literals
-__version__ = '$Id$'
 
 import codecs
 import os
@@ -295,11 +294,9 @@ def create_user_config(args=None, force=False):
                                           config_text=config_text))
 
         pywikibot.output(u"'%s' written." % _fnc)
-    except:
-        try:
+    except BaseException:
+        if os.path.exists(_fnc):
             os.remove(_fnc)
-        except:
-            pass
         raise
